requirements.txt
# Misaka Cipher - Requirements
# Python dependencies for Sprint 1: Nexus Core

# Google AI
google-generativeai>=0.3.0

# OpenAI
openai>=1.0.0

# xAI Grok - using requests for HTTP API
requests>=2.31.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
